AG-6924 - Fix DropShadow option change detection.
