+++ charts-packages/ag-charts-community/src/scene/node.ts
@@ -2,24 +2,13 @@ import { Scene } from "./scene";
 import { Matrix } from "./matrix";
 import { BBox } from "./bbox";
 import { createId } from "../util/id";
-import { windowValue } from "../util/window";
+import { ChangeDetectable, SceneChangeDetection, RedrawType } from './changeDetectable';
+
+export { SceneChangeDetection, RedrawType };
 
 export enum PointerEvents {
     All,
     None
-}
-
-export enum RedrawType {
-    NONE, // No change in rendering.
-
-    // Canvas doesn't need clearing, an incremental re-rerender is sufficient.
-    TRIVIAL, // Non-positional change in rendering.
-
-    // Group needs clearing, a semi-incremental re-render is sufficient.
-    MINOR, // Small change in rendering, potentially affecting other elements in the same group.
-
-    // Canvas needs to be cleared for these redraw types.
-    MAJOR, // Significant change in rendering.
 }
 
 export type RenderContext = {
@@ -35,61 +24,11 @@ export type RenderContext = {
     };
 }
 
-export function SceneChangeDetection(opts?: {
-    redraw?: RedrawType,
-    type?: 'normal' | 'transform' | 'path' | 'font',
-    convertor?: (o: any) => any,
-    changeCb?: (o: any) => any,
-}) {
-    const { redraw = RedrawType.TRIVIAL, type = 'normal', changeCb, convertor } = opts || {};
-    
-    const debug = windowValue('agChartsSceneChangeDetectionDebug') != null;
-
-    return function (target: any, key: string) {
-        // `target` is either a constructor (static member) or prototype (instance member)
-        const privateKey = `__${key}`;
-
-        if (!target[key]) {
-            // Remove all conditional logic from runtime - generate a setter with the exact necessary
-            // steps, as these setters are called a LOT during update cycles.        
-            const setterJs = `
-                ${debug ? 'var setCount = 0;' : ''}
-                function set_${key}(value) {
-                    const oldValue = this.${privateKey};
-                    ${convertor ? 'value = convertor(value);' : ''}
-                    if (value !== oldValue) {
-                        this.${privateKey} = value;
-                        ${debug ? `console.log({ t: this, property: '${key}', oldValue, value, stack: new Error().stack });` : ''}
-                        ${type === 'normal' ? 'this.markDirty(this, ' + redraw + ');' : ''}
-                        ${type === 'transform' ? 'this.markDirtyTransform(' + redraw + ');' : ''}
-                        ${type === 'path' ? `if (!this._dirtyPath) { this._dirtyPath = true; this.markDirty(this, redraw); }` : ''}
-                        ${type === 'font' ? `if (!this._dirtyFont) { this._dirtyFont = true; this.markDirty(this, redraw); }` : ''}
-                        ${changeCb ? 'changeCb(this);' : ''}
-                    }
-                };
-                set_${key};
-            `;
-            const getterJs = `
-                function get_${key}() {
-                    return this.${privateKey};
-                };
-                get_${key};
-            `;
-            Object.defineProperty(target, key, {
-                set: eval(setterJs),
-                get: eval(getterJs),
-                enumerable: true,
-                configurable: false,
-            });
-        }
-    }
-}
-
 /**
  * Abstract scene graph node.
  * Each node can have zero or one parent and belong to zero or one scene.
  */
-export abstract class Node { // Don't confuse with `window.Node`.
+export abstract class Node extends ChangeDetectable { // Don't confuse with `window.Node`.
 
     /**
      * Unique node ID in the form `ClassName-NaturalNumber`.
@@ -518,7 +457,6 @@ export abstract class Node { // Don't confuse with `window.Node`.
         ctx.clearRect(topLeft.x, topLeft.y, bottomRight.x - topLeft.x, bottomRight.y - topLeft.y);
     }
 
-    private _dirty: RedrawType = RedrawType.MAJOR;
     markDirty(_source: Node, type = RedrawType.TRIVIAL, parentType = type) {
         if (this._dirty > type) {
             return;
