+++ charts-packages/ag-charts-community/src/scene/shape/line.ts
@@ -76,6 +76,8 @@ export class Line extends Shape {
 
         this.fillStroke(ctx);
 
+        this.fillShadow?.markClean();
+        this.strokeShadow?.markClean();
         super.render(renderCtx);
     }
 }
